Create additTask method
