+++ src/style.css
@@ -64,3 +64,7 @@ body {
   opacity: 0.5;
   background-color: red;
 }
+
+[contenteditable] {
+  outline: 0px solid transparent;
+}
